EDELIVERY-9465 old ws plugin - allow download of messages that were already downloaded and without changing the status of the message

+++ Plugin-WS/Domibus-default-ws-plugin/src/main/java/eu/domibus/plugin/ws/webservice/deprecated/WebServicePluginImpl.java
@@ -315,14 +315,14 @@ public class WebServicePluginImpl implements BackendInterface {
         }
 
         String trimmedMessageId = messageExtService.cleanMessageIdentifier(retrieveMessageRequest.getMessageID());
+        boolean markAsAcknowledged = retrieveMessageRequest.isMarkAsAcknowledged();
         WSMessageLogEntity wsMessageLogEntity = wsMessageLogDao.findByMessageId(trimmedMessageId);
-        if (wsMessageLogEntity == null) {
+        if(markAsAcknowledged && wsMessageLogEntity == null) {
             LOG.businessError(DomibusMessageCode.BUS_MSG_NOT_FOUND, trimmedMessageId);
             throw new RetrieveMessageFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createFault("No message with id [" + trimmedMessageId + "] pending for download"));
         }
-
         eu.domibus.plugin.ws.generated.header.common.model.org.oasis_open.docs.ebxml_msg.ebms.v3_0.ns.core._200704.UserMessage userMessage =
-                getUserMessage(retrieveMessageRequest, trimmedMessageId);
+                downloadUserMessage(retrieveMessageRequest, trimmedMessageId, markAsAcknowledged);
         eu.domibus.plugin.ws.generated.header.common.model.org.oasis_open.docs.ebxml_msg.ebms.v3_0.ns.core._200704.Messaging messagingWs =
                 WebServiceImpl.EBMS_OBJECT_FACTORY.createMessaging();
         messagingWs.setUserMessage(userMessage);
@@ -335,20 +335,21 @@ public class WebServicePluginImpl implements BackendInterface {
         ebMSHeaderInfo.value = messagingMapper.messagingFromEntity(messagingWs);
 
         try {
-            messageAcknowledgeExtService.acknowledgeMessageDeliveredWithUnsecureLoginAllowed(trimmedMessageId, new Timestamp(System.currentTimeMillis()));
+            messageAcknowledgeExtService.acknowledgeMessageDeliveredWithUnsecureLoginAllowed(trimmedMessageId, new Timestamp(System.currentTimeMillis()), markAsAcknowledged);
         } catch (AuthenticationExtException | MessageAcknowledgeExtException e) {
             //if an error occurs related to the message acknowledgement do not block the download message operation
             LOG.error("Error acknowledging message [" + retrieveMessageRequest.getMessageID() + "]", e);
         }
-
-        // remove downloaded message from the plugin table containing the pending messages
-        wsMessageLogDao.delete(wsMessageLogEntity);
+        if(markAsAcknowledged) {
+            // remove downloaded message from the plugin table containing the pending messages
+            wsMessageLogDao.delete(wsMessageLogEntity);
+        }
     }
 
-    private eu.domibus.plugin.ws.generated.header.common.model.org.oasis_open.docs.ebxml_msg.ebms.v3_0.ns.core._200704.UserMessage getUserMessage(RetrieveMessageRequest retrieveMessageRequest, String trimmedMessageId) throws RetrieveMessageFault {
+    private eu.domibus.plugin.ws.generated.header.common.model.org.oasis_open.docs.ebxml_msg.ebms.v3_0.ns.core._200704.UserMessage downloadUserMessage(RetrieveMessageRequest retrieveMessageRequest, String trimmedMessageId, boolean markAsDownloaded) throws RetrieveMessageFault {
         eu.domibus.plugin.ws.generated.header.common.model.org.oasis_open.docs.ebxml_msg.ebms.v3_0.ns.core._200704.UserMessage userMessage;
         try {
-            userMessage = wsPlugin.downloadMessage(trimmedMessageId, null);
+            userMessage = wsPlugin.downloadMessage(trimmedMessageId, null, markAsDownloaded);
         } catch (final MessageNotFoundException mnfEx) {
             if (LOG.isDebugEnabled()) {
                 LOG.debug(MESSAGE_NOT_FOUND_ID + retrieveMessageRequest.getMessageID() + "]", mnfEx);
